Minor change - fixing a build warning.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@5851 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/current/tests/sourcefitting/RadioSourceTests.h
@@ -216,7 +216,7 @@ namespace askap {
 	  itsGaussSource.defineBox(itsSection,itsFitparams,2); 
 	  float *fluxarray = new float[itsGaussSource.boxSize()];
 	  PixelInfo::Object2D spatMap = itsGaussSource.getSpatialMap();
-	  for (int i = 0; i < itsGaussSource.boxSize(); i++) fluxarray[i] = 0.;
+	  for (size_t i = 0; i < itsGaussSource.boxSize(); i++) fluxarray[i] = 0.;
 	  for(size_t x=0;x<arrayDim;x++) {
 	    for(size_t y=0;y<arrayDim;y++){ 
 	      if (spatMap.isInObject(x, y)) {
